refactor(x64): modernize EXEBlock and PEBuilder for 64-bit pointer safety
- Add GetMachineCodeBlockBytePointer helper to CEXEBlock.
- Replace pointer-to-DWORD truncation casts with uintptr_t across EXEBlock.cpp and PEBuilder.cpp.
- Add test_exeblock_x64.cpp TDD unit test suite.
- Pass 100% of CTest suite (455/455 tests).

+++ DBProCompiler/DBPCompiler/EXEBlock.cpp
@@ -1344,7 +1344,7 @@ bool CEXEBlock::InitDebug(HINSTANCE hInstance, LPVOID pDHookS, LPVOID pDHookJ, L
 					if(m_pDataArray[(d*10)+0]==2)
 					{
 						char* pStr = (char*)m_pDataStringsArray[d];
-						*(DWORD*)&m_pDataSpace[(d*10)+2] = (DWORD)pStr;
+						*(uintptr_t*)&m_pDataSpace[(d*10)+2] = (uintptr_t)pStr;
 					}
 				}
 				g_CORE_PassDataPtrs(m_pDataSpace, m_pDataSpace+m_dwDataSpaceSize);
@@ -1394,7 +1394,7 @@ bool CEXEBlock::InitDebug(HINSTANCE hInstance, LPVOID pDHookS, LPVOID pDHookJ, L
 					if(m_pDataArray[(d*10)+0]==2)
 					{
 						char* pStr = (char*)m_pDataStringsArray[d];
-						*(DWORD*)&m_pDataSpace[(d*10)+2] = (DWORD)pStr;
+						*(uintptr_t*)&m_pDataSpace[(d*10)+2] = (uintptr_t)pStr;
 					}
 				}
